API

Macros

AutoPrettyPrinting.@mime_type — Macro

@mime_type "type"

Registers "type" as a MIME type registered with this package at package definition time. Future calls to @custom_tile and @pprint with no provided mime_types arguments will generate methods corresponding to this type, along with previously declared mime types.

AutoPrettyPrinting.@custom_tile — Macro

@custom_tile [mime_types=nothing] [base_show=false] T => custom_tile_expr

Defines a AutoPrettyPrinting.custom_tile method for type T and optionally provided mime_types using custom_tile_expr as the function's body.

Arguments

mime_types=nothing: The MIME types to generate. Either a String, a vect expression of Strings, or nothing, which will use the package defaults.

base_show::Bool=false: If true, will define Base.show(io::IO, ::MIME(mime), ::T) for each mime in mime_types

custom_tile_expr: Expression to be used for the body of the custom_tile method. Each expression should use a _obj_ placeholder to indicate the object of type T and a _mime_ placeholder to indicate the mime type.

AutoPrettyPrinting.@def_pprint — Macro

@def_pprint [properties=nothing] [mime_types=nothing] [base_show=false] T

Auto generates pretty printing methods for type T and optionally provided mime_types.

Arguments

properties=nothing: Either a Symbol or a vect expression of Symbols corresponding to the properties to use from x::T. Will default to using the fieldnames of x if unspecified

mime_types=nothing: The MIME types to generate. Either a String, a vect expression of Strings, or nothing, which will use the package defaults.

base_show::Bool=false: If true, will define Base.show(io::IO, ::MIME(mime), ::T) for each mime in mime_types

AutoPrettyPrinting.@def_pprint_atomic — Macro

@def_pprint_atomic [to_string=nothing] [mime_types=nothing] T

Registers type T as an atomic type (i.e., one with no subfields nor subelements) and defines custom_tile(x::T, mime::MIME{S}) = to_string(x) for each mime type.

Arguments

to_string=nothing: Function expression that returns the string representation of x. Defaults to Base.string if not provided.

mime_types=nothing: The MIME types to generate. Either a String, a vect expression of Strings, or nothing, which will use the package defaults.

all_mime_types::Bool=false: If true, ignores any provided mime_types and defines custom_tile(x::T, mime::MIME) for a generic MIME input

AutoPrettyPrinting.@hide_typename — Macro

@hide_typename custom_tile(...)

Hides the top-level typename generated by invoking the provided custom_tile expression. Only relevant for top-level types that include their own typename in their custom_tile expression.

Functions

AutoPrettyPrinting.custom_tile — Method

custom_tile(x, mime::MIME; kwargs...)

Returns the layout associated with x for a given mime type.

Defaults to PrettyPrinting.tile(x) if unspecified.

You can override this method for objects x::T of your own custom type T to control how the objects of type T are rendered by this package for a given MIME type.

See also @custom_tile

AutoPrettyPrinting.custom_tile_horiz — Method

custom_tile_horiz(x, mime::MIME; kwargs...)

Returns the horizontal layout associated with x for a given mime type.

Defaults to custom_tile(x, mime; kwargs...) if unspecified.

AutoPrettyPrinting.custom_tile_vert — Method

custom_tile_vert(x, mime::MIME; kwargs...)

Returns the vertical layout associated with x for a given mime type.

Defaults to custom_tile(x, mime; kwargs...) if unspecified.

AutoPrettyPrinting.pprint — Method

pprint(io::IO, mime::MIME, obj)

Pretty prints obj to io according to its custom_tile + mime representation.

AutoPrettyPrinting.repr_pretty — Function

repr_pretty([mime], x; context=nothing)

If istextmime(mime) is true, returns an AbstractString containing the representation of x in the requested mime type rendered with the AutoPrettyPrinting.pprint method.

Otherwise returns repr(mime, x; context).

If mime is not provided, defaults to MIME"text/plain".

Types

AutoPrettyPrinting.KeyValue — Type

KeyValue(key, value)

Represents a generic (key, value) pair, with custom tile representation key = value (as opposed to key => value for the built-in Pair type).

AutoPrettyPrinting.PPrintContext — Type

PPrintContext(io::IO, [mime::MIME])

Create a PPrintContext that wraps a given stream.

Subsequent calls to Base.show(context::PPrintContext, ::MIME, x) will use the pretty printing machinery provided in this package to render x. This type is useful primarily when x has a custom_tile method defined (or provided by this package), but the primary Base.show methods are defined outside of this package and should not be overridden.

Additionally, if a mime::MIME type is provided, subsequent calls to Base.show(context::PPrintContext, x) will dispatch using the provided mime instance.